pyproject.toml
[tool.poetry]
name = "telegram-notifier"
version = "0.1.0"
description = "A simple CLI tool for sending Telegram notifications"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "telegram_notifier"}]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^20.7"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[tool.poetry.scripts]
telegram-notifier = "telegram_notifier.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true